refactor(backend): extract user-permission layer into shared authz module
PermissionResolver (Layer A) is a user-level authorization concern, not a
tool concern — but it lived inside app_tools and AppToolContext owned it, while
handlers checked the same membership/ownership rules through a separate path.

Move PermissionResolver, RepoPermissionResolver, and the Capability/Resource/
ResourceKind vocabulary into a new top-level `authz` module so handlers,
services, and agent tools can share one resolver. app_tools keeps only
AgentPolicy (Layer B, agent scope) and references authz. This makes the RBAC
seam a single point: swap the resolver impl, keep the
(actor, capability, resource) signature.
