+++ backend/src/app_tools/user.rs
@@ -7,7 +7,7 @@ use ailoy::{to_value, tool::ToolDescBuilder, tool_func};
 use serde_json::json;
 
 use crate::app_tools::context::{arg_str, arg_uuid, denied, ok, tool_error, AppToolContext};
-use crate::app_tools::policy::{Capability, Resource};
+use crate::authz::{Capability, Resource};
 use crate::app_tools::AppTool;
 
 pub(crate) fn tools(ctx: &Arc<AppToolContext>) -> Vec<AppTool> {
